# src/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Data API, database migration scripts, and backend services for all Podverse models.",
    "version": "1.0.0",
    "title": "Podverse API",
    "termsOfService": "https://podverse.fm/terms/",
    "contact": {
      "email": "dev@podverse.fm"
    },
    "license": {
      "name": "GNU Affero General Public License v3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "host": "localhost:1234",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/mediaRef": {
      "get": {
        "tags": [
          "mediaRef"
        ],
        "summary": "Find mediaRefs by fields",
        "description": "Returns an array of mediaRefs",
        "operationId": "getMediaRefs",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authors",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "categories",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endTime",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "episodeDuration",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "episodeId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "episodeImageUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "episodeMediaUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "episodePubDate",
            "in": "query",
            "type": "string",
            "format": "date"
          },
          {
            "name": "episodeSummary",
            "in": "query",
            "type": "string"
          },
          {
            "name": "episodeTitle",
            "in": "query",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "podcastFeedUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "podcastGuid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "podcastImageUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "podcastIsExplicit",
            "in": "query",
            "type": "boolean"
          },
          {
            "name": "podcastTitle",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startTime",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/MediaRef"
            }
          }
        }
      },
      "post": {
        "tags": [
          "mediaRef"
        ],
        "summary": "Add a mediaRef",
        "description": "Returns the new mediaRef",
        "operationId": "createMediaRef",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "authors": ["id"],
                "categories": ["id"],
                "description": "string",
                "endTime": "integer",
                "episodeDuration": "integer",
                "episodeGuid": "string",
                "episodeId": "string",
                "episodeImageUrl": "url",
                "episodeLinkUrl": "url",
                "episodeMediaUrl": "url",
                "episodePubDate": "date",
                "episodeSummary": "string",
                "episodeTitle": "string",
                "isPublic": "boolean",
                "ownerId": "string",
                "podcastFeedUrl": "url",
                "podcastGuid": "string",
                "podcastId": "string",
                "podcastImageUrl": "url",
                "podcastIsExplicit": "boolean",
                "podcastTitle": "string",
                "startTime": "integer",
                "title": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/MediaRef"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "mediaRef"
        ],
        "summary": "Update a mediaRef",
        "description": "Returns the updated mediaRef",
        "operationId": "updateMediaRef",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "endTime": "integer",
                "id": "string - required",
                "isPublic": "boolean",
                "startTime": "integer",
                "title": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/MediaRef"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/MediaRef"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/mediaRef/{mediaRefId}": {
      "get": {
        "tags": [
          "mediaRef"
        ],
        "summary": "Find mediaRef by ID",
        "description": "Returns a single mediaRef",
        "operationId": "getMediaRefById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mediaRefId",
            "in": "path",
            "description": "ID of mediaRef to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/MediaRef"
            }
          },
          "404": {
            "description": "MediaRef not found"
          }
        }
      },
      "delete": {
        "tags": [
          "mediaRef"
        ],
        "summary": "Delete mediaRef by ID",
        "description": "Returns success or error message",
        "operationId": "getMediaRefById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mediaRefId",
            "in": "path",
            "description": "ID of mediaRef to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/MediaRef"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "MediaRef not found"
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/playlist": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Find playlists by fields",
        "description": "Returns an array of playlists",
        "operationId": "getPlaylists",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "description",
            "in": "query",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "mediaRefs",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          }
        }
      },
      "post": {
        "tags": [
          "playlist"
        ],
        "summary": "Add a playlist",
        "description": "Returns the new playlist",
        "operationId": "createPlaylist",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "description": "string",
                "isPublic": "boolean",
                "itemsOrder": ["id"],
                "mediaRefs": ["id"],
                "ownerId": "string",
                "title": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "tags": [
          "playlist"
        ],
        "summary": "Update a playlist",
        "description": "Returns the updated playlist",
        "operationId": "updatePlaylist",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "description": "string",
                "id": "string - required",
                "isPublic": "boolean",
                "itemsOrder": [
                  "string"
                ],
                "mediaRefs": [
                  "string"
                ],
                "title": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/playlist/{playlistId}": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Find playlist by ID",
        "description": "Returns a single playlist",
        "operationId": "getPlaylistById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "description": "ID of playlist to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "404": {
            "description": "Playlist not found"
          }
        }
      },
      "delete": {
        "tags": [
          "playlist"
        ],
        "summary": "Delete playlist by ID",
        "description": "Returns success or error message",
        "operationId": "getPlaylistById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "description": "ID of playlist to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Playlist not found"
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "authorization": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Author": {
      "type": "object",
      "required": [
        "id",
        "name",
        "slug"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        }
      }
    },
    "Category": {
      "type": "object",
      "required": [
        "id",
        "slug",
        "title"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        }
      }
    },
    "Episode": {
      "type": "object",
      "required": [
        "id",
        "mediaUrl"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "mediaUrl": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "type": "integer"
        },
        "episodeType": {
          "type": "string"
        },
        "guid": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "isExplicit": {
          "type": "boolean"
        },
        "isPublic": {
          "type": "boolean"
        },
        "linkUrl": {
          "type": "string"
        },
        "mediaFilesize": {
          "type": "integer"
        },
        "mediaType": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "pubDate": {
          "type": "string",
          "format": "date"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "mediaRefs": {
          "$ref": "#/definitions/MediaRef"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "FeedUrl": {
      "type": "object",
      "required": [
        "url",
        "podcast"
      ],
      "properties": {
        "url": {
          "type": "string"
        },
        "isAuthority": {
          "type": "boolean"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "MediaRef": {
      "type": "object",
      "required": [
        "id",
        "episodeMediaUrl",
        "startTime"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "episodeMediaUrl": {
          "type": "string"
        },
        "startTime": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "endTime": {
          "type": "integer"
        },
        "episodeDuration": {
          "type": "integer"
        },
        "episodeGuid": {
          "type": "string"
        },
        "episodeId": {
          "type": "string"
        },
        "episodeImageUrl": {
          "type": "string"
        },
        "episodeLinkUrl": {
          "type": "string"
        },
        "episodePubDate": {
          "type": "string",
          "format": "date"
        },
        "episodeSummary": {
          "type": "string"
        },
        "episodeTitle": {
          "type": "string"
        },
        "isPublic": {
          "type": "boolean"
        },
        "ownerId": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "podcastFeedUrl": {
          "type": "string"
        },
        "podcastGuid": {
          "type": "string"
        },
        "podcastId": {
          "type": "string"
        },
        "podcastImageUrl": {
          "type": "string"
        },
        "podcastIsExplicit": {
          "type": "boolean"
        },
        "podcastTitle": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "episode": {
          "$ref": "#/definitions/Episode"
        },
        "owner": {
          "$ref": "#/definitions/User"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "Playlist": {
      "type": "object",
      "required": [
        "id",
        "owner"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "isMyClips": {
          "type": "boolean"
        },
        "isPublic": {
          "type": "boolean"
        },
        "itemsOrder": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "mediaRefs": {
          "$ref": "#/definitions/MediaRef"
        },
        "owner": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "Podcast": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "feedLastUpdated": {
          "type": "string",
          "format": "date"
        },
        "guid": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "isExplicit": {
          "type": "boolean"
        },
        "language": {
          "type": "string"
        },
        "linkUrl": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "episodes": {
          "$ref": "#/definitions/Episode"
        },
        "feedUrls": {
          "$ref": "#/definitions/FeedUrl"
        }
      }
    },
    "User": {
      "type": "object",
      "required": [
        "id",
        "email"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "subscribedPodcastIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "playlists": {
          "$ref": "#/definitions/Playlist"
        }
      }
    }
  }
}